% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaseControl.R
\name{createCcAnalysesDetails}
\alias{createCcAnalysesDetails}
\title{Create the case-control analyses details}
\usage{
createCcAnalysesDetails(workFolder)
}
\arguments{
\item{workFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}
}
\description{
Create the case-control analyses details
}
\details{
This function creates files specifying the case-control analyses that will be performed.
}
