% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomCovariatesBuilder.R
\name{createCustomCovariatesSettings}
\alias{createCustomCovariatesSettings}
\title{Create custom covariate settings}
\usage{
createCustomCovariatesSettings(useBmi = FALSE, useAlcohol = FALSE,
  useSmoking = FALSE, useDiabetesMedication = FALSE,
  useRheumatoidArthritis = FALSE, useNonRa = FALSE,
  useFatigue = FALSE, useMigraine = FALSE)
}
\arguments{
\item{useBmi}{Create a covariate for BMI (prior to cohort start).}

\item{useAlcohol}{Create a covariate for alcohol use (prior to cohort start).}

\item{useSmoking}{Create a covariate for smoking (prior to cohort start).}

\item{useDiabetesMedication}{Create a covariate for diabetes (medication use) (prior to cohort start).}

\item{useRheumatoidArthritis}{Create a covariate for RA (prior to cohort start).}

\item{useNonRa}{Create a covariate for non-RA, chronic back or chronic neck pain (prior to cohort start).}

\item{useFatigue}{Create a covariate for fatigue or lack of energy (prior to cohort start).}

\item{useMigraine}{Create a covariate for migraine or chronic headache (prior to cohort start).}
}
\description{
Create custom covariate settings
}
