% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsAndTables.R
\name{createPlotsAndTables}
\alias{createPlotsAndTables}
\title{Generate plots and tables}
\usage{
createPlotsAndTables(connectionDetails, cdmDatabaseSchema,
  oracleTempSchema, outputFolder, blind = TRUE)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{outputFolder}{Name of local folder where the results were generated; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{blind}{Blind results? If true, no real effect sizes will be shown. To be used during development.}
}
\description{
Generate plots and tables
}
\details{
Requires that the CohortMethod and CaseControl analyses have been executed
}
