% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantifyingBiasInApapStudies.R
\docType{package}
\name{QuantifyingBiasInApapStudies}
\alias{QuantifyingBiasInApapStudies}
\alias{QuantifyingBiasInApapStudies-package}
\title{QuantifyingBiasInApapStudies}
\description{
QuantifyingBiasInApapStudies
}
