% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{generateReport}
\alias{generateReport}
\title{Generate a report containing the main results}
\usage{
generateReport(outputFolder)
}
\arguments{
\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}
}
\description{
Generate a report containing the main results
}
